Initial import.
ADT_API/ is the Alta 0 and level 1 API.
AltaSetup/ is a program from the examples that is being refactored into our control program.
rdAltaUDP/ is a test UDP reader program that parses the data and outputs it to screen.

AltaSetup & rdAltaUDP will work together to test the device.
AltaSetup will also work with nidas.